pyproject.toml
[tool.poetry]
name = "yeast_cc_db"
version = "0.0.0"
description = "A django managed database intended to enable the ananalysis of Scerevisiae Calling Cards data"
authors = ["chase mateusiak <chasem@wustl.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "^3.6.0"
asttokens = "^2.2.1"
backcall = "^0.2.0"
blessings = "^1.7"
boto3 = "^1.26.88"
botocore = "^1.29.88"
click = "^8.1.3"
coverage = "^7.2.1"
decorator = "^5.1.1"
dj-database-url = "^1.2.0"
Django = "^4.1.7"
django-configurations = "^2.4"
django-filter = "^22.1"
django-model-utils = "^4.3.1"
django-storages = "^1.13.2"
django-unique-upload = "^0.2.1"
djangorestframework = "^3.14.0"
executing = "^1.2.0"
factory-boy = "^3.2.1"
Faker = "^17.6.0"
flake8 = "^6.0.0"
ghp-import = "^2.1.0"
gunicorn = "^20.1.0"
importlib-metadata = "^6.0.0"
ipdb = "^0.13.13"
ipython = "^8.11.0"
jedi = "^0.18.2"
Jinja2 = "^3.1.2"
jmespath = "^1.0.1"
MarkupSafe = "^2.1.2"
matplotlib-inline = "^0.1.6"
mccabe = "^0.7.0"
mergedeep = "^1.3.4"
mock = "^5.0.1"
newrelic = "^8.7.0"
packaging = "^23.0"
parso = "^0.8.3"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
prompt-toolkit = "^3.0.38"
psycopg2-binary = "^2.9.5"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.2"
pycodestyle = "^2.10.0"
pyflakes = "^3.0.1"
Pygments = "^2.14.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
pytz = "^2022.7.1"
PyYAML = "^6.0"
pyyaml_env_tag = "^0.1"
s3transfer = "^0.6.0"
six = "^1.16.0"
sqlparse = "^0.4.3"
stack-data = "^0.6.2"
tomli = "^2.0.1"
traitlets = "^5.9.0"
urllib3 = "^1.26.14"
watchdog = "^2.3.1"
wcwidth = "^0.2.6"
zipp = "^3.15.0"
mkdocs = "^1.4.2"
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.8.1"
pandas = "^2.0.0"
numpy = "^1.24.2"
celery = "^5.2.7"
redis = "^4.5.4"
flower = "^1.2.0"
drf-spectacular = "^0.26.2"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
pytest-django = "^4.5.2"
ipykernel = "^6.21.3"


[tool.poetry.group.dev.dependencies]
pytest-runner = "^6.0.0"
pytest-cov = "^4.0.0"
autopep8 = "^2.0.2"
pydotplus = "^2.0.2"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
